Final levels, tweaked and fixes

+++ Game/Assets/_Game/Scripts/Shop/Items/JumpForceUpgradeItem.cs
@@ -3,6 +3,6 @@
 namespace Swen.Shop.Items {
   [CreateAssetMenu(menuName = "Shop/Jump Force Upgrade Item")]
   public class JumpForceUpgradeItem : ShopItem {
-    public int JumpForceAddition;
+    public float JumpForceAddition;
   }
 }
