+++ Game/Assets/_Game/Scripts/Shop/Items/MovementSpeedUpgradeItem.cs
@@ -3,6 +3,6 @@
 namespace Swen.Shop.Items {
   [CreateAssetMenu(menuName = "Shop/Movement Speed Upgrade Item")]
   public class MovementSpeedUpgradeItem : ShopItem {
-    public int MovementSpeedAddition;
+    public float MovementSpeedAddition;
   }
 }
